The Swamp-code was kinda broken. It didn't kill drowning players correctly. :-)
I fixed that and added a big extra saving-throw
for having the woodsman skill. It's very unlikely
now to die despite of being a woodsman.

(In case someone doesn't know: Players die in swamp
if they stand still for too long. It's a direct kill,
no immunities can save the player.)
--AV