I had to re-adjust some of the recently changed levels in pupland monsters. As maintainer of the pupland set I hope it is okay when I do this, have written on the list about it too.
Some monsters just need those levels to be able to
hit and damage high level players.
Evil Masters are a somewhat special case:
I have moved most all attributes out of the map into
the arches (So the arches contain the "real version").
The reason for this is I fear the fallback to
arch-values when these monsters happen to apply equipment.

--AndreasV