Experimental code to attempt a supply and demand based economy for shops.  This code lays the framework for a supply and demand based pricing scheme, and tracks all purchases and sales by players.  It calculates what the price for the item should be, based on the data it has collected.
However.. the code to make it actually chaneg the prices in the game has
not been added.  At this stage, the code only tracks the prices in a
database, var/supplydb.  It is my intention to collect this data from
running servers to see what effect such a change would make.  It is quite
possible that this scheme will be backed out, and a different one used
based on the data collected.  The data should give us a reasonable
picture of what is sold and purchased in the game, to better judge future
economic changes.

There should be no gameplay changes from this commit.