Clean exit
Added window closed events checks so the client exits when a window is closed.
Works with cfclient and gtkclient. In normal mode, when main window is closed,
exit the client. In split mode, closing any window will exit the program. This
behaviour is much better than a pack of assert error or a program still working
without any window to manage.