New utility to build wiki monster pages.
This is my first commit of my bwp (build wiki pages) program.  The
program works at the moment, except for a sorting issue, but I thought
it'd be easier for anyone who wants to help fix that if the files were
on svn.  I plan to add additional features, and welcome suggestions.
Command to compile (on FreeBSD, anyway) is in the comments at the head
of the file.