Prevent core dumps when maps are not found.
Previously, the server would continue running if the regions file was
not found, causing it to try to load the initial map. If the maps
directory was not present, both steps would fail, which leads to a call
to abort() and causes a core dump.

We can fix this by checking the return value of init_regions() and
exiting with a failure if the maps are not found.