Remove manual output socket buffer handling
Our manual buffer code doesn't handle partial writes. While logging in a
client, this sometimes causes the server to skip the tail of a partial
write, resulting in corrupt data being sent to the client.

Instead of handling this issue ourselves, grow the OS output socket
buffer to SOCKETBUFSIZE and disconnect the client if this buffer ever
gets filled, consistent with current behavior, only now it also works
during login.