Revert "Fix bug #809 -- Server: configure doesn't check for flex."
configure does check for lex, and sets it to ':' if it can't find it, so
that end users of the source distribution do not need lex installed in
order to build the server.

Also prefer AM_PROG_LEX over AC_PROG_LEX, because it uses the missing
script on systems that do not have lex.