Make find_insert_pos() much faster
When loading a map, this function used to take O(n^2). It searched
through all objects on a tile until it found the top. If INS_MAP_LOAD is
set, it throws away the result and uses the invariant GET_MAP_TOP. If
INS_ABOVE_FLOOR_ONLY is set, it throws that result away and uses the
floor it found earlier.

During map load, it is really INS_ON_TOP that matters. In that case,
check for that flag first and return immediately using the invariant,
without doing the O(n) search. Otherwise, do the search and return the
topmost floor or the topmost free space.

Since players like to put their things in large piles, this change
drastically reduces the time taken to load maps like player apartments.